+++ src/player/CurveNode.h
@@ -49,11 +49,11 @@ class AVG_API CurveNode : public VectorNode
         const glm::vec2& getPos4() const;
         void setPos4(const glm::vec2& pt);
 
-        double getTexCoord1() const;
-        void setTexCoord1(double tc);
+        float getTexCoord1() const;
+        void setTexCoord1(float tc);
 
-        double getTexCoord2() const;
-        void setTexCoord2(double tc);
+        float getTexCoord2() const;
+        void setTexCoord2(float tc);
 
         virtual void calcVertexes(VertexArrayPtr& pVertexArray, Pixel32 color);
 
@@ -65,8 +65,8 @@ class AVG_API CurveNode : public VectorNode
         glm::vec2 m_P2;
         glm::vec2 m_P3;
         glm::vec2 m_P4;
-        double m_TC1;
-        double m_TC2;
+        float m_TC1;
+        float m_TC2;
 
         std::vector<glm::vec2> m_LeftCurve;
         std::vector<glm::vec2> m_RightCurve;
